Character Main Screen refactoring
