+++ RickAndMortyAPI/ui/screens/character/main/CharacterMainScreen.swift
@@ -107,53 +107,3 @@ struct CharacterMainScreen: AppView {
 #Preview {
     CharacterMainScreen(uiComponent: RootComponent().uiComponent)
 }
-
-class CharacterMainViewModel : ViewModel<CharacterMainScreenState> {
-    init() {
-        super.init(viewState: CharacterMainScreenState())
-    }
-    
-    func showCharacterDetail(by characterDetail: CharacterDetailUI) {
-        mutate { viewState in
-            viewState.setCurrentScreen(
-                screenType: CharacterMainScreenState.ScreenType.Detail(
-                    characterDetail: characterDetail
-                )
-            )
-        }
-    }
-    
-    func showCharacterList() {
-        mutate { viewState in
-            viewState.setCurrentScreen(
-                screenType: CharacterMainScreenState.ScreenType.List
-            )
-        }
-    }
-}
-
-struct CharacterMainScreenState : ViewState {
-    internal var errorMessage: FailureError?
-    internal var showProgressView: Bool = false
-    var currentScreen: ScreenType = ScreenType.List
-    
-    enum ScreenType {
-        case List
-        case Detail(characterDetail: CharacterDetailUI)
-        
-        func isDetail() -> Bool {
-            switch self {
-            case .List: false
-            case .Detail(_): true
-            }
-        }
-    }
-}
-
-extension CharacterMainScreenState {
-    mutating func setCurrentScreen(screenType: ScreenType) {
-        currentScreen = screenType
-    }
-}
-
-
